Create README.md
The World Bank is a great source of current and historic data for every country in the world. The purpose of this map is to easily access each country's "World Bank profile" page. By clicking on a country, the page opens in a new tab. Some of the data include population, GDP, CO2 emissions, and life expectancy. I used Leaflet and CartoDB tiles to create the map.